Breaking down cartesian into solve_cartesian code, removing dependence on sunpy map
