+++ pfsspy/utils.py
@@ -251,44 +251,3 @@ def car_to_cea(m, method='interp'):
         meta_out[key] = header_out[key]
     m_out = sunpy.map.Map(data_out, header_out)
     return m_out
-
-def get_units(mapobj):
-    return mapobj.meta['cunit1'],mapobj.meta['cunit1']
-def Mm_to_unit(unit):
-    """Convert Megameters to degree or arcsec.
-        1arcsec = 725 km on the Sun. 1 degree= 3600 arcsec
-
-    Args:
-        unit (string): "degree" or "arcsec"
-
-    Returns:
-        (float): Conversion factor FROM Mm TO unit
-    """
-    if unit=='deg':
-        return 1000.0/(725*3600)
-    elif unit=='arcsec':
-        return 1000.0/(725)
-    else:
-        raise ValueError("Units should be degree or arcsec.")
-def calc_cosine_heliocentric_angle(map_ob):
-    """Calculates the heliocentric cosine angle given a map object.
-      The cosine value, called "mu" is defined as :
-        mu = cos(theta), where
-        
-        sin(theta) = sqrt(x**2+y**2)/R, 
-        
-        where x,y are longitude and latitude resp in arcsec, and R = 1919/2 = 959.6 arcsec.
-        
-        By simple trignometry, we get:
-        cos(theta) = sqrt(R**2-x**2-y**2)/R
-        
-        Reference: arXiv:1307.2410v1 
-
-    Args:
-        map_ob (sunpy.map.GenericMap): a sunpy map object
-    """
-    R=1919.0/2
-    latitude = np.linspace(map_ob.bottom_left_coord.data.lat,map_ob.top_right_coord.data.lat,map_ob.data.shape[0]).reshape([-1,1]).value
-    longitude = np.linspace(map_ob.bottom_left_coord.data.lon,map_ob.top_right_coord.data.lon,map_ob.data.shape[1]).reshape([1,-1]).value
-    mu = np.sqrt(R**2-np.square(latitude)-np.square(longitude))/R
-    return mu
